PR Feedback, fixed model export
